[upd] - refactor nurses and users livewire component using traits
